% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DealGPL570}
\alias{DealGPL570}
\title{This function deals GPL570 (Affymetrix Human Genome U133 Plus 2.0 Array) RAW.tar file using the robust multi-array average expression measure.}
\usage{
DealGPL570(file, type = "probeIDAndGeneSymbol")
}
\arguments{
\item{file}{a file name specified by either a variable of mode character, or a double-quoted string, which is of 'GSE*_RAW.tar' which is of platform GPL570.}

\item{type}{a character string indicating which type of data frame is to be outputted. One of 'probeIDAndGeneSymbol'(default), 'probeID', or 'geneSymbol'.}
}
\value{
a data frame which is a expression profile.
}
\description{
This function deals GPL570 (Affymetrix Human Genome U133 Plus 2.0 Array) RAW.tar file using the robust multi-array average expression measure.
}
\examples{

file <- system.file("extdata", "GSE104683_RAW.tar", package = "DealGPL570")
file
#Next step would run for about 15s, so you can try it yourself or view the vignettes
\dontrun{
result <- DealGPL570(file = file)
}
}
